Add blog plot generation script with real compilation results
- generate_blog_plots.py: runs compile + optimize + benchmark pipeline
- Normal model: 38.5x faster than PyTensor (0.06 vs 2.46 us/eval)
- LinReg model: 7.1x faster (0.49 vs 3.50 us/eval)
- Hierarchical model: 8.8x faster (0.57 vs 5.04 us/eval)
- Optimization progress plots (autoresearcher-style)
- Waterfall and timeline plots for both models
- Benchmark comparison bar chart

+++ compiled_models/normal/src/bench.rs
@@ -1,14 +1,15 @@
-use pymc_compiled_model::GeneratedLogp;
+
+use pymc_compiled_model::generated::GeneratedLogp;
 use nuts_rs::CpuLogpFunc;
-use std::time::Instant;
 use std::io::{self, BufRead};
+use std::time::Instant;
 
 fn main() {
     let stdin = io::stdin();
     let mut lines = stdin.lock().lines();
 
     // First line: number of iterations
-    let n_iters: u64 = lines.next().unwrap().unwrap().trim().parse().unwrap();
+    let n_iters: usize = lines.next().unwrap().unwrap().trim().parse().unwrap();
 
     // Second line: parameter vector
     let param_line = lines.next().unwrap().unwrap();
